+++ app/app/src/main/java/kia/app/entry/MainActivity.java
@@ -1,6 +1,7 @@
 package kia.app.entry;
 
 import android.Manifest;
+import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.BroadcastReceiver;
@@ -523,6 +524,7 @@ public final class MainActivity extends Activity {
     }
 
     @Override
+    @SuppressLint("UnspecifiedRegisterReceiverFlag")
     protected void onResume() {
         super.onResume();
         activityVisible = true;